+++ webpack.config.js
@@ -2,8 +2,11 @@
  * webpack 打包配置
  * 2018年1月4日 星期四
  */
+const fs = require('fs')
 const webpack = require('webpack')
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin')           // 文档
+const Pkg = require('./package.json')
+
 // 压缩插件
 const nodeEnv = process.env.NODE_ENV 
 
@@ -11,11 +14,17 @@ class Queue{
     constructor(){
         this.FileList = []
     }
-    js(name){
+    /**
+     * @param {string} name 源文件名
+     * @param {string|null} refName 编译以后的文件名
+     * @param {function|null} 回调函数 
+     */
+    js(name, refName, callback){
+        refName = refName? refName:name
         var opt = {
             entry: './src/'+name+'.js',
             output: {
-                filename: './build/'+name + 
+                filename: './build/'+refName + 
                     (nodeEnv !== 'production'? '':'.min')+
                     '.js'
             }
@@ -45,6 +54,9 @@ class Queue{
                 })
             )
         }
+        if('function' == typeof callback){
+            callback(opt)
+        }
         this.FileList.push(opt)
         return this
     }
@@ -53,8 +65,31 @@ class Queue{
     }
 }
 
+// 编译时时间运行
+(function(){
+    var _json = {
+        version: Pkg.version,
+        release: Pkg.release,
+        author: Pkg.author
+    }
+    fs.writeFileSync('./version.js', `export default ${JSON.stringify(_json)}`)
+})()
+
 module.exports = new Queue()
     .js('worker')
+
+    .js('worker', 'workflow')   // 别名编译， 实际用于 zonmaker 框架的工作流库
+
+    .js('browser.workerflow', 'workflow')   // 别名编译， 实际用于 zonmaker 框架的工作流库  
+
+    // 别名编译， 实际用于 zonmaker 框架的工作流库  AMD
+    .js('browser.workerflow.umd', 'workflow.umd', (opt) => {
+        opt.output.libraryTarget = 'umd'
+        opt.externals = {
+            'raphael': 'raphael'
+        }
+    })
+
     .js('tree')
     .js('surong')   // index
 .getFiles()